# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Privy Stream API
  version: 1.0.0
  description: Внешний REST API Gateway для аудио-стримингового сервиса
servers:
  - url: http://localhost:8080
paths:
...
  /refresh:
    post:
      summary: Обновить пару access/refresh токенов
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Токены обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '400':
          description: Invalid request data
        '401':
          description: Invalid or expired refresh token
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
...